requirement.txt
pypdf
transformers
einops
accelerate
langchain
bitsandbytes
sentence_transformers
llama-index-llms-huggingface
llama_index
huggingface_hub
torch
python-dotenv